# packages/api-contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthBodyEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          },
          "429": {
            "description": "Rate limit exceeded; see Retry-After",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "summary": "Exchange credentials for a token pair",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rotated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthBodyEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          },
          "429": {
            "description": "Rate limit exceeded; see Retry-After",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "summary": "Rotate a refresh token for a new pair",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset code sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordEnvelopeDto"
                }
              }
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          },
          "429": {
            "description": "Rate limit exceeded; see Retry-After",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "summary": "Request password reset OTP via email",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageEnvelopeDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/problem+json": {}
            }
          },
          "404": {
            "description": "No such resource, or the caller may not know it exists",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          },
          "429": {
            "description": "Rate limit exceeded; see Retry-After",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "summary": "Reset password using OTP code",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Change password for currently authenticated user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Session revoked"
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "summary": "Revoke the current login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll_v1",
        "parameters": [],
        "responses": {
          "204": {
            "description": "All sessions revoked"
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Revoke every login of the current user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "The currently authenticated user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me/abilities": {
      "get": {
        "operationId": "AuthController_abilities_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Packed CASL rules",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbilitiesEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "The caller's permissions as packed CASL rules, for the client to build its ability",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Starter App API",
    "description": "Success responses are wrapped in `{ data, meta? }`; errors follow RFC 9457 Problem Details with `application/problem+json`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the access token only. Swagger adds the Bearer prefix."
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 256
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 256
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthBodyEnvelopeDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string"
              },
              "user": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "fullName": {
                    "type": "string"
                  },
                  "tenantId": {
                    "type": "string"
                  },
                  "scope": {
                    "type": "string",
                    "enum": [
                      "platform",
                      "tenant"
                    ]
                  },
                  "roles": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "key",
                        "name"
                      ]
                    }
                  }
                },
                "required": [
                  "id",
                  "email",
                  "fullName",
                  "scope",
                  "roles"
                ]
              },
              "refreshToken": {
                "type": "string"
              },
              "csrfToken": {
                "type": "string"
              }
            },
            "required": [
              "accessToken",
              "user"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 1
          }
        },
        "default": {}
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "ForgotPasswordEnvelopeDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "expiresInSeconds": {
                "type": "integer"
              }
            },
            "required": [
              "message",
              "expiresInSeconds"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "otp": {
            "type": "string",
            "pattern": "^\\d{6}$"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "maxLength": 256
          }
        },
        "required": [
          "email",
          "otp",
          "newPassword"
        ]
      },
      "MessageEnvelopeDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            },
            "required": [
              "message"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "minLength": 1
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "maxLength": 256
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "PublicUserEnvelopeDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "fullName": {
                "type": "string"
              },
              "tenantId": {
                "type": "string"
              },
              "scope": {
                "type": "string",
                "enum": [
                  "platform",
                  "tenant"
                ]
              },
              "roles": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "key",
                    "name"
                  ]
                }
              }
            },
            "required": [
              "id",
              "email",
              "fullName",
              "scope",
              "roles"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "AbilitiesEnvelopeDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "rules": {
                "type": "array",
                "items": {}
              }
            },
            "required": [
              "rules"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
...
    }
  }
}